Refactor(package) : 패키지 구조 변경
패키지를 아키텍처와 도메인 영역으로 분리
