+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/architecture/controller/ItemSearchController.java
@@ -1,0 +1,13 @@
+package kr.ac.hongik.dsc2023.ydy.team1.core.architecture.controller;
+
+import kr.ac.hongik.dsc2023.ydy.team1.core.architecture.dto.request.SearchItemRequestDTO;
+import kr.ac.hongik.dsc2023.ydy.team1.core.architecture.dto.response.ResponseWrapper;
+import kr.ac.hongik.dsc2023.ydy.team1.core.architecture.dto.response.SearchItemResponseDTO;
+import org.springframework.http.ResponseEntity;
+
+/**
+ * Item 검색 API 를 정의하는 Controller 인터페이성
+ */
+public interface ItemSearchController {
+    ResponseWrapper<ResponseEntity<SearchItemResponseDTO>> search(SearchItemRequestDTO requestDTO);
+}
