+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/architecture/dto/response/ItemCreateResponseDTO.java
@@ -1,0 +1,7 @@
+package kr.ac.hongik.dsc2023.ydy.team1.core.architecture.dto.response;
+
+/**
+ * Item 생성 응답을 표현하는 추상클래스
+ */
+public abstract class ItemCreateResponseDTO {
+}
