--- src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/domain/entity/Item.java
@@ -1,20 +1,0 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.domain.entity;
-
-import lombok.Getter;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "item")
-@Getter
-public class Item {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "price", nullable = false)
-    private int price;
-}
