--- src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/dto/request/konbini/KonbiniBrand.java
@@ -1,5 +1,0 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.dto.request.konbini;
-
-public enum KonbiniBrand {
-    GS25,CU,SEVENELEVEN,EMART24
-}
