--- src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/dto/request/konbini/KonbiniPromotion.java
@@ -1,5 +1,0 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.dto.request.konbini;
-
-public enum KonbiniPromotion {
-    ONE_PLUS_ONE,TWO_PLUS_ONE
-}
