--- src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/dto/request/port/SearchStrength.java
@@ -1,8 +1,0 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.dto.request.port;
-
-/**
- * 상품 검색시, 검색 키워드와 검색 결과물의 일치 정도를 결정할때 사용하는 클래스
- */
-public enum SearchStrength {
-    STRONG
-}
