--- src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/dto/response/konbini/KonbiniSearchItemResponseDTO.java
@@ -1,10 +1,0 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.dto.response.konbini;
-
-import kr.ac.hongik.dsc2023.ydy.team1.core.dto.response.port.SearchItemResponseDTO;
-import lombok.Getter;
-import lombok.experimental.SuperBuilder;
-
-@SuperBuilder
-@Getter
-public class KonbiniSearchItemResponseDTO extends SearchItemResponseDTO {
-}
