--- src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/dto/response/port/ItemCreateResponseDTO.java
@@ -1,7 +1,0 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.dto.response.port;
-
-/**
- * Item 생성 응답을 표현하는 추상클래스
- */
-public abstract class ItemCreateResponseDTO {
-}
