--- src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/dto/response/port/ResponseWrapper.java
@@ -1,14 +1,0 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.dto.response.port;
-
-import lombok.Getter;
-import lombok.experimental.SuperBuilder;
-
-/**
- * 각종 API의 응답을 감싸는 랩퍼 클래스
- * @param <T>
- */
-@SuperBuilder
-@Getter
-public abstract class ResponseWrapper<T>{
-    protected T response;
-}
