+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/konbini/dto/response/KonbiniSearchItemResponseDTO.java
@@ -1,0 +1,10 @@
+package kr.ac.hongik.dsc2023.ydy.team1.core.konbini.dto.response;
+
+import kr.ac.hongik.dsc2023.ydy.team1.core.architecture.dto.response.SearchItemResponseDTO;
+import lombok.Getter;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+@Getter
+public class KonbiniSearchItemResponseDTO extends SearchItemResponseDTO {
+}
