+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/konbini/model/KonbiniBrand.java
@@ -1,0 +1,5 @@
+package kr.ac.hongik.dsc2023.ydy.team1.core.konbini.model;
+
+public enum KonbiniBrand {
+    GS25,CU,SEVENELEVEN,EMART24
+}
