+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/konbini/model/KonbiniPromotion.java
@@ -1,0 +1,5 @@
+package kr.ac.hongik.dsc2023.ydy.team1.core.konbini.model;
+
+public enum KonbiniPromotion {
+    ONE_PLUS_ONE,TWO_PLUS_ONE
+}
